Add files via upload
In this project, I utilize data from four prominent companies, namely Apple, Google, Microsoft, and Amazon. In Part 1, I conduct an analysis of various aspects of their stocks. This includes examining the closing prices of each stock, evaluating the total volume of stocks traded on a daily basis, investigating the daily price changes in the stocks, and calculating the monthly and yearly mean of the closing feature.